requirements.txt
# ============================================
# MOTOR DE COBRANZA - Dependencias
# ============================================

# ========== FastAPI & Server ==========
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
python-multipart==0.0.17

# ========== Flask (para módulos legacy) ==========
flask==3.0.3
flask-cors==4.0.1

# ========== HTTP & WebSockets ==========
httpx==0.27.2
websockets==12.0
requests==2.32.3

# ========== Twilio ==========
twilio==9.2.3

# ========== OpenAI & LangChain (VERSIONES COMPATIBLES) ==========
openai==1.54.3
langchain-openai==0.2.8
langchain-core==0.3.13

# ========== Base de Datos ==========
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# ========== Utilidades ==========
pydantic==2.9.2
pydantic-settings==2.6.1